backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
psycopg-pool
httpx
python-dotenv
